+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  * Added Story object
  * Added Giveaway object
  * Added GiveawayWinners object
+ * Added ExternalReplyInfo object
 
 ### Changed
  * Refactored the way HTTP requests are handled & methods are called